Move code interpreting a carbon-copy of a sent message's body.
The body of a sent message's copy needs to be written to the
corresponding message window manually since libpurple does not
know what to do with those.

Previously, this was done right away, which caused some problems
when a message did not contain a body, or there was additional
data to be interpreted later.

The new implementation gives other plugins a chance to handle the
message as well, before handling it a second time to work around
libpurple.

This fixes #21.